docs: `ApplicantCard` 주석 추가, 파일명 변경
